fix: code clean up, fix for straight, straight flush and royal flush
This will clean up some code from the Poker class and improve poker test
feature.

Added a fix for checking for Straight hand, we didn't check for the
sequence of cards before - only checked for the diff of first and last
card. This was inefficient and did not check if the immediate card was a
sequence.