fix: update the deck and poker lib
This will update the way we handle the deck and poker class.

For Poker class, we now put all checks in one single if-else statement
to compute the winner. Also pass the preShuffle option via the
constructor. Finally, add a tie breaker total for high card comparision,
in the event both player and dealer have the same high card. We now
calculate the total of all other cards to break that high card tie.

For Deck class, we now use one single suit array to fill the deck.
Remove uniqueID and use an internal ID iterator. Use shift() method to
remove cards from the deck, instead of performing a splice(). Adjust the
parse() function to parse the suit properly.